limbs_buffer: introduce support for native endian multiprecision integers
Currently there is support for multiprecision integers in either big or
little endian layout. In particular for intermediate computations, it is
highly desirable to use an architecture's native representation for
performance reasons.

Define a layout for this case, the "native endian" format. It stores the limbs
from least significant to most significant order, with each limb in native
endian format.

Note that unlike it is the case with the exisiting big and little endian
layouts, the representation of "partial" high limbs would be undefined: think
of the case of a partial high limb, last in memory order, on big endian
architectures. Additional rules could get introduced to make this case
well-defined, but it would only clutter the code and be diametral against the
native endian layout's purpose of improving performance. So don't support
partial high limbs for the native endian format and require that all backing
byte slice buffers' length are aligned to a multiple of the limb size.

Signed-off-by: Nicolai Stange <nstange@suse.de>