Don't zeroize local variables
Currently there are a couple of places in the code which would wrap
certain local variables with zeroize::Zeroizing<>. However, this by
itself doesn't give any guarantees whatsoever, as copies of said data
can leak out in any uncontrolled manner. What is really required would
be some compiler directive instructing it to clear out the stack and
registers upon function return, which we don't have.

Not only does wrapping locals in Zeroizing<> not provide any guarantees, it
also tends to make the compiler produce worse code, because it e.g.
prevents it to scalarize aggregates etc.

So don't do it.

Signed-off-by: Nicolai Stange <nstange@suse.de>