limb: provide a black_box_l() for LimbType values
Rust's core::hint::black_box() incurrs some non-negligible cost for the
generated code:
- first of all, probably due to it being generic and supporting types of
  arbitrary size, it moves everything through the stack,
- furthermore, supposedly because it's aiming primarily at benchmarks,
  it prevents the compiler from optimizing unused values away if passed
  through black_box().

In addition to that, the documentation says
  "[...] it must not be relied upon to control critical program behavior.
   This immediately precludes any direct use of this function for cryptographic
   or security purposes."
I would argue that in its current implementation, this statement just does not
apply (well, except for runtime-optimized ILs like WASM perhaps), but it might
in the future.

Implement an alternative, black_box_l(), specific to the case of LimbType values
only, but more suitable for our use-case to just make the compiler forget any
knowledge it has inferred about the given value up to the point of the
invocation.

Signed-off-by: Nicolai Stange <nstange@suse.de>