limb: provide a LimbChoice implementation replacing subtle
The current strategy to use the cond_choice_to_mask() and
cond_select_with_mask() in order to save some of subtle's
conditional_select() invocations make the code more and more cumbersome
to read.

Provide a LimbChoice, loosely analogous to to subtle's Choice, but for
selecting between LimbType alternatives only.

Signed-off-by: Nicolai Stange <nstange@suse.de>