shift: make tests to exercise bitwise shifts
Now that mp_lshift_mp() and mp_rshift_mp() have been improved to support
not only shift distances which are a multiple of bytes, but arbitrary
distances, it would be great to have the tests cover that.

Make the shift implementation's tests to exercise shift distances which are not
a multiple of bytes.

Signed-off-by: Nicolai Stange <nstange@suse.de>