shift: make mp_{lshift,rshift}_mp() constant-time
The current mp_lshift_mp() and mp_rshift_mp() implementations' runtime
depends on the shift distance. This is pretty much inherent to the
problem at hand, because for e.g. lshift, there's nothing to shift in
for all the bits below the shift distance.

However, to still obliterate the shift distance in terms of the number of
instructions executed, make both lshift and rshift to wrap around and fill the
tail or head ranges respectively corresponding to the shifted in zeroes with
contents (garbage) from the multiprecision integer's opposite end before
eventually clearing them out in constant-time.

Memory access patterns for shift distances exceeding a limb width, or a
cacheline even, would still manifest themselves in side-effects observable
through side-channels though, but there's little than can be (reasonably) done
about it. Also, a common application of e.g. rshift is to divide a given
integer by powers of two and it can be assumed that in cryptographic contexts,
where sensitive integers are random, such an integer's count of factors of two
are (way) below the limb width with high likelihood.

Signed-off-by: Nicolai Stange <nstange@suse.de>