limb: introduce ct_lsb_mask_l()
The generation of LimbType masks with a specified number of least
significant bits all set is a common task, yet it is a bit cumbersome to
implement, at least if the case of all LIMB_BITS bits being set is to be
supported. Implement it once so it can be used from everywhere needed.

Signed-off-by: Nicolai Stange <nstange@suse.de>