limb: provide x86_64 specific alternative implementations of add and sub
x86_64's add and sub instructions provide direct access to the carry
state of the respective operation, which is more efficient than that
high bit examination of the the result and the input operands the
software implementation is doing, of course. Make the x86_64 add and sub
instructions available as drop-in substitutes for the corresponding software
primitives. Just as it's the case for mul and div already, this is gated
under the Cargo "enable_arch_math_asm" feature.

Signed-off-by: Nicolai Stange <nstange@suse.de>