limb: provide x86_64 specific alternative implementations of is_{,non}_zero()
The software implementations of ct_is_zero_l() and ct_is_nonzero_l()
respectively exploit the fact that for non-zero value v, for either v or
-v the most significant bit would be non-zero. Testing for this involves
a couple of instructions, of course, like a logical negation and an add
for the arithmetic negation, a logical or and a shift. On x86_64, the
same can be achieved with only two instructions: a 'test' and a
subsequent 'setnz'.

Provide x86_64 specific drop-in substitures for ct_is_zero_l() and
ct_is_nonzero_l(), gated by the "enable_arch_math_asm" Cargo feature
just as it's the case for add, sub, mul and div already.

Signed-off-by: Nicolai Stange <nstange@suse.de>