invmod: implement inversion modulo generic moduli
The currently available modulo inversion primitive based on the Extended Euclid
Algorithm, mp_ct_inv_mod_odd(), supports only odd moduli.

It is fairly straight-forward to build further on that and support inversion
modulo arbitrary moduli. Let n = n_{*} * 2^e be the modulus such that n_{*} is
odd. Let v be an element whose inverse modulo n is to be computed. Obtain its
inverses modulo n_{*} and also, modulo 2^e. The latter can be done efficiently
using Hensel lifting. The two intermediate results are then combined to
an inverse modulo n using Garner's algorithm.

For the constant-time Hensel lifting, note that for any E >= e, an inverse
modulo 2^E can easily get turned into one module 2^e by clearing all bits
beyond e. So, to achieve constant time, always compute the inverses
modulo 2^E, with E set to the worst case maximum, i.e. the modulus' bit width.
With that, the total number of multiplications required is quadratic in the
number of the modulus' limbs, with a constant factor of 1/3. Note that
this procedure needs to get run twice: once for the input value to invert and
once more for n_{*}.

The subsequent application of Garner's method involves two more multi-precision
multiplications, which, together with the ones from the Hensel step,
amounts to a total of 2 + 2 * 1/3 times the number of the modulus' limbs
squared of multiplications.

Signed-off-by: Nicolai Stange <nstange@suse.de>