euclid: implement GCD for generic arguments
The currently available GCD implementation, ct_gcd_odd_mp_mp(), has the
limitation that the first operand must be odd.

Build on it to implement a GCD primitive working for generic operands,
no matter if even or not: count common powers of two, rshift both
operands by that distance and invoke ct_gcd_odd_mp_mp(), possibly after
swapping the shifted operands to pass the odd one as the first argument.

Finally, make the LCM implementation to use this new primitive instead
of open-coding it.

Signed-off-by: Nicolai Stange <nstange@suse.de>