div: provide modulo convenience wrappers around div
The multiprecision primitives, ct_div_mp_mp(), ct_div_mp_l() and
specializations take an optional Option<> output argument for recveiving
the resulting quotient. Often however, callers are only used in the
remainder. If the quotient output argument is set to None, the
primitives are a bit cumbersome to use, because the associated generic
type parameter needs to get set to some dummy MpIntMutByteSlice
implementation.

Provide convience wrappers for the modulo-only operations.

Signed-off-by: Nicolai Stange <nstange@suse.de>