limbs_buffer: provide shrink_to() primitive for the multiprecision integers
For users of the library, it is sometimes desirable to virtually shrink some
multiprecision integer bytes slice whenever it is known that its high part is
all zeroes. For example, this can save time in subsequent computations.

Implement a ::shrink_to() for all the MP integer byte slice wrappers. It
will shrink the given slice to the specified length, or rather to the
minimum >= that value compatible with the respective slice layout's length
alignment constraints.

Signed-off-by: Nicolai Stange <nstange@suse.de>