limb: implement a non-constant time limb divisor
The one-time setup of a constant-time limb divisor, i.e. of a
CtLDivisor, requires the division of a certain value by the given LimbType
divisor to determine the associated runtime-constant multiplier. In certain
cases however, like for the prime wheel, the divisor value is not sensitive at
all, only the dividends are. As this one-time setup constant-time division is
fairly expensive -- it's a bitwise long division -- it would be good to have a
non-constant-time, but more efficient alternative for this use-case.

To enable non-constant-time construction of CtLDivisor for non-sensitive
divisor value, implement non-constant-time double-limb-by-limb division. Expose
it by means of the new struct NonCtLDivisor.

As the basic double-limb-by-limb division primitive required only needs to
support cases where the quotient fits a limb, restrict the implementation
accordingly. This is in complete analogy to CtLDivisor, which exposes its
division functionality only through a private trait, CtLDivisorPrivate, to
prevent accidental misuse by external users. Rename CtLDivisorPrivate to
"LDivisorPrivate" and make the new NonCtLDivisor implement it.

In fact, two alternative implementations of NonCtLDivisor are provided:
a generic software implementation and another one mapping directly to
x86_64's "div" instruction.

Finally, extend CtLDivisor by a non-constant-time constructor,
CtLDivisor::nonct_new(), implemented on the grounds of the new NonCtLivisor.

Signed-off-by: Nicolai Stange <nstange@suse.de>