limbs_buffer: allow for out-of-range test_bit() invocations
A common task is to check whether some integer is odd. For this, it first
needs to get verified that it is not empty and only afterwards can
its least significant bit get tested.

Enable MpIntByteSliceCommon::test_bit() to test bits at positions not
actually covered by the backing byte slice and make it return "zero" in
this case.

Signed-off-by: Nicolai Stange <nstange@suse.de>