Improve overall error checking and reporting
For many functions, especially in the montgomery, euclid and invmod
submodules, the length constraints on the argument's input/output
multiprecision integer byte slices are non-trivial, meaning they could
potentially get wrong by users of the library. As it currently stands, these
constraints are not enforced at all, but only checked in debug builds by means
of corresponding debug_assert!() statements.

It is however not exactly ideal to rely on Rust's built-in array bounds
checker, which should be considered more as a last resort safety net,
because that would simply crash the application upon OOB accesses. It
would certainly be favorable to enable applications to recover
gracefully and also, to provide some meaningful error information beyond
a mere stacktrace to where the OOB access had happened.

Improve on the current situation: introduce proper function argument checks as
is appropriate and return corresponding Result::Err()s upon failure.

Signed-off-by: Nicolai Stange <nstange@suse.de>