euclid, invmod: improve corner case error reporting of ct_inv_mod{,_odd}_mp_mp()
Both, ct_inv_mod_odd_mp_mp() and ct_inv_mod_mp_mp() behave well for the
corner cases of the value to invert being equal to zero or the modulus
equalling one even in that they would return a fixed result of zero then.
However, the inverse of zero does not exist and for a modulus of one, the
entire multiplicative group doesn't even exist.

So these cases indiciate that something fishy is going and it is
definitely better to alert the user of the library by reporting an
error.

Do that, make ct_inv_mod_odd_mp_mp() and ct_inv_mod_mp_mp() return an
error of OperandsNotCoprime in case the value to invert is zero (which
includes the modulus == 1 case).

Signed-off-by: Nicolai Stange <nstange@suse.de>