limbs_buffer: prepare for arbitrary backing slice element types
The current code is throughout based on the assumption that
multiprecision integers are always backed by an &[u8] (or &mut [u8])
slice. However, the native endian slice type is bound to get converted
to be backed by a slice of LimbTypes for performance reasons. Prepare
for this by making the the traits in mod limbs_buffer generic over the
slice type and adapt the rest of the code as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>