limbs_buffer: provide MpMutUInt::copy_from_cond()
There is already an unconditional copy_from() primitive available in
trait MpMutUint. In preparation of dropping the conditional support
for Montgomery multiplications, provide a suitable alternative for
exponentiation algorithms and alike: implement
MpMutUInt::copy_from_cond() so that the (unconditional) product can get
overwritten with one of the operands again if the given condition is
false.

Signed-off-by: Nicolai Stange <nstange@suse.de>