div: implement support for decremented divisors
For sampling of non-zero, random integers with a given upper bound, the
method of oversampling is a common procedure: a random integer with
some excess bits (e.g. 64) is sampled from a random number generator,
then subsequently reduced modulo the given upper bound minus one and
the finally incremented by one to make it non-zero.

In order to avoid an extra buffer allocation just for computing the
divisor of the upper bound minus one, make the division primitives
support LimbType decrements to be applied "on the fly" to the divisor.

Signed-off-by: Nicolai Stange <nstange@suse.de>